Move interval to utils and split utils in module
